// src/api/generated/model/getRoomPromoCodeUsagesResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { PromoCodeUsageItemDto } from './promoCodeUsageItemDto';

export interface GetRoomPromoCodeUsagesResponseDto {
  /** List of promo code usages on the current page */
  items: PromoCodeUsageItemDto[];
  /** Current page number (1-based) */
  page: number;
  /** Number of items per page */
  size: number;
  /** Total number of items across all pages */
  total: number;
  /** Total number of pages */
  totalPages: number;
}